+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added access to the CloudBrigadierManager from Brigadier-enabled command managers
  - Added parameter injectors
  - Store currently parsing command argument in the command context
+ 
+### Fixed
+ - Use the correct default range for Double and Float parsers in the StandardParserRegistry
 
 ## [1.1.0] - 2020-10-24